backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
groq==0.11.0
langgraph==0.2.34
langchain-core==0.3.10
langchain-groq==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1
python-dateutil==2.9.0
